+++ package.json
@@ -4,8 +4,10 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "rm -rf dist && tsc && cp -R public/** dist && node dist/app.js",
-    "lint": "gts lint"
+    "build": "rm -rf dist && tsc && cp -R public/** dist",
+    "start": "npm run build && node dist/app.js",
+    "test": "mocha -r ts-node/register src/**/*.spec.ts",
+    "lint": "gts lint src"
   },
   "author": "Google LLC",
   "license": "Apache-2.0",
@@ -18,8 +20,12 @@
   },
   "devDependencies": {
     "@types/express": "^4.17.1",
+    "@types/mocha": "^10.0.1",
+    "chai": "^4.3.7",
     "gts": "^3.1.1",
+    "mocha": "^10.2.0",
     "prettier": "^2.8.3",
+    "ts-node": "^10.9.1",
     "typescript": "^4.9.4"
   }
 }
